pyproject.toml
[project]
name = "create-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-qgis>=4.0.1",
]
